+++ packages/core/src/Transfer/types.js
@@ -1,0 +1,12 @@
+/**
+ * These are instances of react components,
+ * both built-in and custom ones.
+ *
+ * @typedef {Object} ReactElement
+ */
+
+/**
+ * @typedef {Object} Option
+ * @prop {string} label
+ * @prop {string} value
+ */